move functions into helpers
this makes them availiable to other functions that might need. e.g. array_slicer should live in generic but also needs to create sliders
